pyproject.toml
[tool.poetry]
name = "crm-backend"
version = "0.1.0"
description = ""
authors = ["Nakadzima <pascal.german@yandex.ru>"]
readme = "README.md"
packages = [{include = "crm_backend"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "3.2.16"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
python-decouple = "^3.7"
pillow = "^9.4.0"
django-cors-headers = "^3.13.0"
drf-yasg = "^1.21.4"
whitenoise = "^6.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
